+++ src/main/java/com/sbproject/piclocket/PicLocketApplication.java
@@ -6,8 +6,10 @@ import com.sbproject.piclocket.security.JwtProperties;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication
+@EnableScheduling
 @EnableConfigurationProperties({
         AwsProperties.class, JwtProperties.class, CorsProperties.class
 })